fix(llma): Use float instead of int for errorCount in LLMTrace schema
The TypeScript schema defines errorCount as `number` which the schema
builder correctly converts to Python `float`. Update Python schema to
match the expected type.

+++ posthog/schema.py
@@ -4282,7 +4282,7 @@ class LLMTrace(BaseModel):
     )
     aiSessionId: Optional[str] = None
     createdAt: str
-    errorCount: Optional[int] = None
+    errorCount: Optional[float] = None
     events: list[LLMTraceEvent]
     id: str
     inputCost: Optional[float] = None
